feat: ThemeKey validation 로직 변경 및 ThemeTemplate로 이동
